requirements.txt
pyvisa
pyvisa-py
PyUSB
flask
flask_restful
flask-jsonpify
jsonify